// src/generated/server/worldmonitor/emergency/v1/service_server.ts
export interface Alert {
  id: string;
  event: string;
  severity: string;
  urgency: string;
  headline: string;
  description: string;
  instruction: string;
  areaDesc: string;
  onset: string;
  expires: string;
  polygon: string;
}

export interface ListActiveAlertsRequest {
  island: string;
}

export interface ListActiveAlertsResponse {
  alerts: Alert[];
}